chore: move subscribe to channel.rs file
- move subcribe function to project tree under mod channels
- change the type of parameter to i32 to match the channel_id type
