+++ src/channels/routes/channels.rs
@@ -1,9 +1,16 @@
 use diesel::prelude::*;
-
 use diesel_async::RunQueryDsl;
-use rocket::{serde::json::Json, State};
+use rocket::{
+    response::stream::{Event, EventStream},
+    serde::json::Json,
+    tokio::{
+        select,
+        sync::broadcast::{self, error::RecvError, Sender},
+    },
+    Shutdown, State,
+};
 
-use crate::{Channel, DbPool, ErrorResponse};
+use crate::{AppState, Channel, DbPool, ChannelMessage, ErrorResponse};
 
 #[get("/<channel_id>")]
 pub async fn get_channel(pool: &State<DbPool>, channel_id: i32) -> Result<Json<Channel>, Json<ErrorResponse>> {
@@ -26,4 +33,44 @@ pub async fn get_channel(pool: &State<DbPool>, channel_id: i32) -> Result<Json<C
         .map(|e| e.into())
         .map_err(|a| a.into())
 
+}
+
+#[get("/<channel_id>/subscribe")]
+pub async fn subscribe(
+    channel_id: i32,
+    sessions: &State<AppState>,
+    mut end: Shutdown,
+) -> EventStream![] {
+    let mut sessions = sessions.clients.lock().await;
+    // Does the sessions exists ?
+    // If not we create it and store in the state (meaning new conversation)
+    let new_cli = if let Some(existing_cli) = sessions.get(&channel_id) {
+        existing_cli.clone()
+    } else {
+        let new_cli = broadcast::channel(10).0;
+        sessions.insert(channel_id, new_cli.clone());
+        new_cli
+    };
+
+    let queue: &Sender<ChannelMessage> = &new_cli;
+
+    // Add the client to the broadcast channel so that he can receive messages
+    let mut client = queue.subscribe();
+
+    EventStream! {
+        // Messages received down the channel are sent to the clients that
+        // they've subscribed to
+        loop {
+            let msg = select! {
+                msg  = client.recv() => match msg {
+                    Ok(msg) => msg,
+                    Err(RecvError::Closed) => break,
+                    Err(RecvError::Lagged(_)) => continue,
+                },
+                _ = &mut end => break,
+            };
+
+            yield Event::json(&msg);
+        }
+    }
 }
