+++ src/channels/routes/mod.rs
@@ -8,6 +8,7 @@ pub fn routes() -> Vec<Route> {
 
     routes.extend(routes![messages::get_channel_history]);
     routes.extend(routes![channels::get_channel]);
+    routes.extend(routes![channels::subscribe]);
 
     routes
 }
